+++ packages/web-api/post-scans-func/index.ts
@@ -3,9 +3,9 @@
 import 'reflect-metadata';
 
 import { Context } from '@azure/functions';
-import { ControllerDispatcher } from '../src/controllerDispatcher';
+import { ControllerDispatcher } from '../src/controller-dispatcher';
 import { ScanRequestController } from '../src/controllers/scan-request-controller';
-import { setupIoContainer } from '../src/setupIoContainer';
+import { setupIoContainer } from '../src/setup-ioc-container';
 
 export async function run(context: Context): Promise<void> {
     const dispatcher = new ControllerDispatcher(ScanRequestController, context, setupIoContainer());
